Update 4_analysis_SRS.md
There was a typo in the equation, calling an empty cell (D24) instead of D2 for the mean to be subtracted in the variance equation. There was also a typo in the cells from column C that were called in that equation. The error is in the linked spreadsheet too.